separate test server and running of lighthouse tests, then use test server for e2e tests as well
