+++ tests/e2e/training-page-test.js
@@ -1,20 +1,21 @@
+import createServer from '../createServer'
 import { Selector, ClientFunction } from 'testcafe'
 require('dotenv').config()
 
-const { GATSBY_ENVIRONMENT } = process.env
-
-const baseUrl =
-  GATSBY_ENVIRONMENT === 'preview'
-    ? process.env.DEPLOY_PRIME_URL
-    : process.env.BASE_URL
-
+const baseUrl = `localhost:3001`
 const trainingPageUrl = `${baseUrl}/training`
 const getWindowLocation = ClientFunction(() => window.location)
 
+let server
 let firstModalLink
-fixture`Training page`.page`${trainingPageUrl}`.beforeEach(async t => {
-  firstModalLink = await Selector('a[data-testid^="modal"]').nth(0)
-})
+fixture`Training page`.page`${trainingPageUrl}`
+  .before(async t => {
+    server = createServer()
+  })
+  .beforeEach(async t => {
+    firstModalLink = await Selector('a[data-testid^="modal"]').nth(0)
+  })
+  .after(() => server.close())
 
 test('should open the correct training course modal when accessed via a link', async t => {
   const coursePathName = await firstModalLink
