+++ tests/lighthouse/launchChromeAndRunLighthouse.js
@@ -1,0 +1,17 @@
+const createServer = require('../createServer')
+const lighthouse = require('lighthouse')
+const chromeLauncher = require('chrome-launcher')
+const config = require('./lh-config.js')
+
+const launchChromeAndRunLighthouse = url => {
+  const server = createServer()
+
+  return chromeLauncher.launch({ chromeFlags: [] }).then(chrome => {
+    return lighthouse(url, { port: chrome.port }, config).then(results => {
+      server.close()
+      return chrome.kill().then(() => results)
+    })
+  })
+}
+
+module.exports = launchChromeAndRunLighthouse
