+++ tests/lighthouse/lighthouse.test.js
@@ -1,4 +1,4 @@
-const serve = require('./createServer.js')
+const launchChromeAndRunLighthouse = require('./launchChromeAndRunLighthouse.js')
 
 jest.setTimeout(60000)
 
@@ -22,102 +22,112 @@ const auditTest = (audits, name, type, number) => {
 }
 
 test('Mobile Homepage', () => {
-  return serve(`http://localhost:3001`).then(({ lhr: { audits } }) => {
-    // https://developers.google.com/web/tools/lighthouse/audits/has-viewport-meta-tag
-    auditTest(audits, 'viewport', 'value')
-    // https://developers.google.com/web/tools/lighthouse/audits/font-sizes
-    auditTest(audits, 'font-size', 'value')
-  })
+  return launchChromeAndRunLighthouse(`http://localhost:3001`).then(
+    ({ lhr: { audits } }) => {
+      // https://developers.google.com/web/tools/lighthouse/audits/has-viewport-meta-tag
+      auditTest(audits, 'viewport', 'value')
+      // https://developers.google.com/web/tools/lighthouse/audits/font-sizes
+      auditTest(audits, 'font-size', 'value')
+    }
+  )
 })
 
 test('SEO', () => {
-  return serve(`http://localhost:3001`).then(({ lhr: { audits } }) => {
-    auditTest(audits, 'meta-description', 'value')
-    auditTest(audits, 'document-title', 'value')
-    auditTest(audits, 'http-status-code', 'value')
-    auditTest(audits, 'is-crawlable', 'value')
-    auditTest(audits, 'robots-txt', 'value')
-    // https://developers.google.com/web/tools/lighthouse/audits/hreflang
-    auditTest(audits, 'hreflang')
-  })
+  return launchChromeAndRunLighthouse(`http://localhost:3001`).then(
+    ({ lhr: { audits } }) => {
+      auditTest(audits, 'meta-description', 'value')
+      auditTest(audits, 'document-title', 'value')
+      auditTest(audits, 'http-status-code', 'value')
+      auditTest(audits, 'is-crawlable', 'value')
+      auditTest(audits, 'robots-txt', 'value')
+      // https://developers.google.com/web/tools/lighthouse/audits/hreflang
+      auditTest(audits, 'hreflang')
+    }
+  )
 })
 
 test('Security', () => {
-  return serve(`http://localhost:3001`).then(({ lhr: { audits } }) => {
-    // https://developers.google.com/web/tools/lighthouse/audits/vulnerabilities
-    auditTest(audits, 'no-vulnerable-libraries')
-    // (https://www.chromestatus.com/features#deprecated
-    auditTest(audits, 'deprecations')
-    // https://developers.google.com/web/tools/lighthouse/audits/noopener
-    auditTest(audits, 'external-anchors-use-rel-noopener', 'size', 1)
-  })
+  return launchChromeAndRunLighthouse(`http://localhost:3001`).then(
+    ({ lhr: { audits } }) => {
+      // https://developers.google.com/web/tools/lighthouse/audits/vulnerabilities
+      auditTest(audits, 'no-vulnerable-libraries')
+      // (https://www.chromestatus.com/features#deprecated
+      auditTest(audits, 'deprecations')
+      // https://developers.google.com/web/tools/lighthouse/audits/noopener
+      auditTest(audits, 'external-anchors-use-rel-noopener', 'size', 1)
+    }
+  )
 })
 
 test('Performance', () => {
-  return serve(`http://localhost:3001`).then(({ lhr: { audits } }) => {
-    auditTest(audits, 'dom-size', 'smaller', 700)
-    auditTest(audits, 'network-requests', 'smaller', 70)
-    auditTest(audits, 'network-requests', 'smaller', 70)
-    auditTest(audits, 'bootup-time', 'smaller', 2600) //  0.89 -- prev 1333
-    auditTest(audits, 'interactive', 'smaller', 10000) //  0.45 -- prev 7788
-    auditTest(audits, 'speed-index', 'smaller', 6000) //  0.71 -- prev 4582
-    auditTest(audits, 'first-contentful-paint', 'smaller', 4700)
-    auditTest(audits, 'first-meaningful-paint', 'smaller', 4700)
-    // https://developers.google.com/web/tools/lighthouse/audits/preload
-    auditTest(audits, 'uses-rel-preload')
-    // https://developers.google.com/web/tools/lighthouse/audits/blocking-resources
-    auditTest(audits, 'render-blocking-resources', 'size', 1) // currently we have one blocking CSS resource (google fonts)
-    // https://developers.google.com/web/tools/lighthouse/audits/minify-css
-    auditTest(audits, 'unminified-css')
-    // https://developers.google.com/speed/docs/insights/MinifyResources
-    auditTest(audits, 'unminified-javascript')
-    // https://developers.google.com/web/tools/lighthouse/audits/optimize-images
-    auditTest(audits, 'uses-optimized-images')
-    // https://developers.google.com/web/tools/lighthouse/audits/oversized-images
-    auditTest(audits, 'uses-responsive-images')
-    // https://developers.google.com/web/tools/lighthouse/audits/redirects
-    auditTest(audits, 'redirects')
-    // https://developers.google.com/web/tools/lighthouse/audits/aspect-ratio
-    auditTest(audits, 'image-aspect-ratio')
-    // https://developers.google.com/web/tools/lighthouse/audits/unused-css-rules
-    auditTest(audits, 'unused-css-rules', 'size', 1)
-  })
+  return launchChromeAndRunLighthouse(`http://localhost:3001`).then(
+    ({ lhr: { audits } }) => {
+      auditTest(audits, 'dom-size', 'smaller', 700)
+      auditTest(audits, 'network-requests', 'smaller', 70)
+      auditTest(audits, 'network-requests', 'smaller', 70)
+      auditTest(audits, 'bootup-time', 'smaller', 2600) //  0.89 -- prev 1333
+      auditTest(audits, 'interactive', 'smaller', 10000) //  0.45 -- prev 7788
+      auditTest(audits, 'speed-index', 'smaller', 6000) //  0.71 -- prev 4582
+      auditTest(audits, 'first-contentful-paint', 'smaller', 4700)
+      auditTest(audits, 'first-meaningful-paint', 'smaller', 4700)
+      // https://developers.google.com/web/tools/lighthouse/audits/preload
+      auditTest(audits, 'uses-rel-preload')
+      // https://developers.google.com/web/tools/lighthouse/audits/blocking-resources
+      auditTest(audits, 'render-blocking-resources', 'size', 1) // currently we have one blocking CSS resource (google fonts)
+      // https://developers.google.com/web/tools/lighthouse/audits/minify-css
+      auditTest(audits, 'unminified-css')
+      // https://developers.google.com/speed/docs/insights/MinifyResources
+      auditTest(audits, 'unminified-javascript')
+      // https://developers.google.com/web/tools/lighthouse/audits/optimize-images
+      auditTest(audits, 'uses-optimized-images')
+      // https://developers.google.com/web/tools/lighthouse/audits/oversized-images
+      auditTest(audits, 'uses-responsive-images')
+      // https://developers.google.com/web/tools/lighthouse/audits/redirects
+      auditTest(audits, 'redirects')
+      // https://developers.google.com/web/tools/lighthouse/audits/aspect-ratio
+      auditTest(audits, 'image-aspect-ratio')
+      // https://developers.google.com/web/tools/lighthouse/audits/unused-css-rules
+      auditTest(audits, 'unused-css-rules', 'size', 1)
+    }
+  )
 })
 
 test('A11y Homepage', () => {
-  return serve(`http://localhost:3001`).then(({ lhr: { audits } }) => {
-    // https://dequeuniversity.com/rules/axe/2.2/image-alt?application=lighthouse
-    auditTest(audits, 'image-alt')
-    // https://dequeuniversity.com/rules/axe/2.2/input-image-alt?application=lighthouse
-    auditTest(audits, 'input-image-alt', 'value')
-    // https://dequeuniversity.com/rules/axe/2.2/label?application=lighthouse
-    auditTest(audits, 'label', 'value')
-    // https://dequeuniversity.com/rules/axe/2.2/link-name?application=lighthouse
-    auditTest(audits, 'link-name')
-    // https://dequeuniversity.com/rules/axe/2.2/list?application=lighthouse
-    auditTest(audits, 'list')
-    // https://dequeuniversity.com/rules/axe/2.2/meta-viewport?application=lighthouse
-    auditTest(audits, 'meta-viewport')
-    // https://dequeuniversity.com/rules/axe/2.2/tabindex?application=lighthouse
-    auditTest(audits, 'tabindex')
-    // https://dequeuniversity.com/rules/axe/2.2/html-lang?application=lighthouse
-    auditTest(audits, 'html-has-lang', 'value')
-    // https://dequeuniversity.com/rules/axe/2.2/valid-lang?application=lighthouse
-    auditTest(audits, 'html-lang-valid', 'value')
-    // https://dequeuniversity.com/rules/axe/2.2/bypass?application=lighthouse
-    auditTest(audits, 'bypass', 'value')
-    // https://dequeuniversity.com/rules/axe/2.2/button-name?application=lighthouse
-    auditTest(audits, 'button-name')
-    // https://dequeuniversity.com/rules/axe/2.2/aria-valid-attr-value?application=lighthouse
-    auditTest(audits, 'aria-valid-attr-value')
-    // https://dequeuniversity.com/rules/axe/2.2/aria-valid-attr?application=lighthouse
-    auditTest(audits, 'aria-valid-attr')
-    // https://dequeuniversity.com/rules/axe/2.2/aria-roles?application=lighthouse
-    auditTest(audits, 'aria-roles')
-    // https://dequeuniversity.com/rules/axe/2.2/aria-allowed-attr?application=lighthouse
-    auditTest(audits, 'aria-allowed-attr')
+  return launchChromeAndRunLighthouse(`http://localhost:3001`).then(
+    ({ lhr: { audits } }) => {
+      // https://dequeuniversity.com/rules/axe/2.2/image-alt?application=lighthouse
+      auditTest(audits, 'image-alt')
+      // https://dequeuniversity.com/rules/axe/2.2/input-image-alt?application=lighthouse
+      auditTest(audits, 'input-image-alt', 'value')
+      // https://dequeuniversity.com/rules/axe/2.2/label?application=lighthouse
+      auditTest(audits, 'label', 'value')
+      // https://dequeuniversity.com/rules/axe/2.2/link-name?application=lighthouse
+      auditTest(audits, 'link-name')
+      // https://dequeuniversity.com/rules/axe/2.2/list?application=lighthouse
+      auditTest(audits, 'list')
+      // https://dequeuniversity.com/rules/axe/2.2/meta-viewport?application=lighthouse
+      auditTest(audits, 'meta-viewport')
+      // https://dequeuniversity.com/rules/axe/2.2/tabindex?application=lighthouse
+      auditTest(audits, 'tabindex')
+      // https://dequeuniversity.com/rules/axe/2.2/html-lang?application=lighthouse
+      auditTest(audits, 'html-has-lang', 'value')
+      // https://dequeuniversity.com/rules/axe/2.2/valid-lang?application=lighthouse
+      auditTest(audits, 'html-lang-valid', 'value')
+      // https://dequeuniversity.com/rules/axe/2.2/bypass?application=lighthouse
+      auditTest(audits, 'bypass', 'value')
+      // https://dequeuniversity.com/rules/axe/2.2/button-name?application=lighthouse
+      auditTest(audits, 'button-name')
+      // https://dequeuniversity.com/rules/axe/2.2/aria-valid-attr-value?application=lighthouse
+      auditTest(audits, 'aria-valid-attr-value')
+      // https://dequeuniversity.com/rules/axe/2.2/aria-valid-attr?application=lighthouse
+      auditTest(audits, 'aria-valid-attr')
+      // https://dequeuniversity.com/rules/axe/2.2/aria-roles?application=lighthouse
+      auditTest(audits, 'aria-roles')
+      // https://dequeuniversity.com/rules/axe/2.2/aria-allowed-attr?application=lighthouse
+      auditTest(audits, 'aria-allowed-attr')
 
-    // https://dequeuniversity.com/rules/axe/2.2/color-contrast?application=lighthouse
-    // auditTest(audits,'color-contrast')
-  })
+      // https://dequeuniversity.com/rules/axe/2.2/color-contrast?application=lighthouse
+      // auditTest(audits,'color-contrast')
+    }
+  )
 })
